+++ src/ribasim_nl/ribasim_nl/__init__.py
@@ -1,7 +1,8 @@
 __version__ = "0.0.1"
 
 from ribasim_nl.cloud import CloudStorage
+from ribasim_nl.model import Model
 from ribasim_nl.network import Network
 from ribasim_nl.reset_index import reset_index
 
-__all__ = ["CloudStorage", "Network", "reset_index"]
+__all__ = ["CloudStorage", "Network", "reset_index", "Model"]
